pyproject.toml
[project]
name = "genai-article-processor"
version = "0.0.0"
description = "Building an Application for Processing Old Article Images with AI."
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "rjoydip", email = "joydipand@gmail.com"}
]
dependencies = [
    "google-genai>=1.12.1",
    "lxml>=5.4.0",
    "python-dotenv",
]
[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pre-commit>=4.2.0",
]
lint = [
    "ruff>=0.11.6",
    "pyright>=1.1.399",
]

[build-system]
requires = ["setuptools", "pip"]
build-backend = "setuptools.build_meta"

# Setup tools
[tool.setuptools.packages.find]
where = ["."]
include = []
exclude = []
namespaces = false

# Pytest
[tool.pytest.ini_options]
 minversion = "6.0"
 addopts = "-ra"
 testpaths = [
    "tests/*",
    "integration",
]
python_files = "test_*.py"

# Logging
[tool.logging]
version = 1

[tool.logging.formatters.formatter]
class = "logging.Formatter"
format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"

[tool.logging.handlers.fh]
level = "DEBUG"
class = "logging.FileHandler"
formatter = "formatter"

[tool.logging.handlers.ch]
level = "ERROR"
class = "logging.StreamHandler"
formatter = "formatter"

[tool.logging.spam_application]
level = "DEBUG"
handlers = ["fh", "ch"]

[tool.logging.spam_application.auxiliary]
propagate = false

[tool.logging.spam_application.auxiliary.Auxiliary]
propagate = false

# UV
[tool.uv]
upgrade = false
default-groups = "all"

# Pyright
[tool.pyright]
venvPath = "."
venv = ".venv"
ignore = ["docs/", "tests/test_*.py"]

# Ruff
[tool.ruff]
exclude = [
    ".git",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "dist",
    "site-packages",
    "venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
